changed GetFiles to run second match only for host
push(@{$files},grep(/^$str$/,(keys %{$logs->{usage}})));
That match should only be if someone is looking for a value from a
single host.  Something like bb10-56-mar/cpu.1m.  Because of that there
is no reason to run it if we're looking for a link.  Saves about 300000
regex calls on my network.