version number bump
I think I'm basically done with code cleanup and making things work,
like 64 bit counters.  Next may be to recreate things like isis2fig or
to work on usage2dia, rrdtool, graphviz, or a mongo backend for parts of it.

Or I may need to modularize more of it to make it more backend agnostic.

Or I may drop it the whole thing for a while because it's working for
me. :)