Fix for line thickness segfaulting with modern GD
For unknown reasons, the brush creation in SetStyle was causing GD to
segfault.  GD has the ability to set line thickness directly so I don't
think the brushes are needed.

It may be that GD always had that feature and Pete wrote his own to
workaround some limitation of it.  I know AdjustForThickness() is needed
to prevent thickened lines from running inside of boxes, but I don't
know if there are any other places setThickness might cause issues.