checking in with broken imgmap code
Checking this in for reference.  The comments are in the source as to
how I might want to fix it.

We're doing some semi-lowlevel dia crap to get this to work.  Arguably
it should be written into a standalone Dia class so others could make
use of it, but I don't feel like I'm doing enough with it to make it
worth anyone elses time.

If we start doing more dia then I may fork that off into it's own
module.