d3 experimental charts
Most of the chart code can be replaced with third party libraries like
D3 that handle things like figuring out date ranges and scaling for you.

One downside of doing this with D3 would be that sending data to the
client with JSON (uncompressed at least) is slower than just sending a
graph of the compiled values.  I may figure out a way to overcome this,
or find a way to use D3 on the server end and send completed svg/png
files.