got CPU line working
I also changed the date_line to black because that is what Chart has.  I
thought it was red for some reason earlier.

I have one or two graphing concerns to solve, then I just need to label
everything and the graph is done and could replace chart in a normal
install.