not happy with the way the graph works now
committing what I've got.

We can't inline them as XML/SVG images.  We can't really call it as JS
library and include it as a script src.

There needs to be more work done to abstract it and I want to get away
from passing around so many variables.  We should be able to just do 1-2
GET's and be done.  Possibly 3 because we're having javascript fetch
the data to render the graph.

Right now it's map page to chart page to /graph to /x/data