d3chart updates
This is broken but has some good ideas.  I'm going to back out of most
of it because the scope is too much for me to bother with right now, but
incorporate the things that are working and make sense.

1.  Added file protection so users can't do .. or / path names.  Need to
check if things like //etc/passwd could still work.
2.  Switching to numbered files instead of named so we could have more
than 3 lines on the graph.  This caused me tons of issues because I
don't know what is going on what data manipulation in javascript.
Honestly we would be better served by skipping all this and working on
NVD3 or some other charting library that makes D3 easy to use.

I still need to add protection for, and allow you to choose which net
you're on.  Something that can check the cibhrc file and figure out if
that network is defined before calling the eval?