Removing +01 from the time division
I think +.01 was to prevent division by zero, but
it doesn't work.
      int(1-1+0.01)=0.

Because the time values are integers and +.01 is less than 5 it will
always round down, so it seems to be pointless.

I checked old versions and it looks like it was always inside the int()
so it shouldn't have worked before either.