removed int() from bdiv
The bad news is if the two numbers aren't integers then bdiv doesn't
work.  The good news is that epoch time minus epoch time should be an
integer.

If we add +.01 to it then either we need int() to convert it back, or
we need to switch to floating point or something.