Mojolicious version of cgi-bin/chart
This covers most of the problems.  I think the only issue with the new
code is clicking on a graph via the imgmap won't zoom to that section,
but instead loads the section with a blank sample 24 hours into the
past.

I think this is caused by stop, span or start not being used properly
but I need to look at the old code to see if I can figure out what
they were doing there.