version bump to 2.0.0
This code had not been tested on anything that was beyond Mojolicious 5.xx. I'm now starting the process of upgrading to conform to the modern version.  I'm setting required versions for Mojolicious and perl

convert to signatures

Fix can't use hash as reference error in d3chart.  This may have never worked and been committed in error?